refactor(components): product components in module
