+++ hooks/useGetItem.js
@@ -2,13 +2,18 @@ import React from "react";
 import { get } from "../api";
 
 const useGetItem = (id) => {
-  const [item, setItem] = React.useState([]);
+  const [item, setItem] = React.useState();
 
-  React.useEffect(() => {
-    get(`/api/products/${id}`).then((res) => setItem(res.data.product));
-  }, [id]);
+  const getProduct = async () => {
+    const resp = await get(`/api/products/${id}`);
+    setItem(resp.data.product);
+  };
 
-  return [item, setItem];
+  React.useEffect(getProduct(), [id]);
+
+  console.log(item);
+
+  return { item, setItem };
 };
 
 export default useGetItem;
