+++ modules/products/components/details/actions/index.jsx
