+++ modules/products/components/details/styled.jsx
