--- modules/products/details/index.jsx
@@ -1,16 +1,0 @@
-import CardProduct from "../../../components/global/Products/Details";
-import ScrollToTop from "../../../hooks/components/ScrollToTop";
-import useGetItem from "../../../hooks/useGetItem";
-import { Container } from "./styled";
-
-function DeatilsMod({ id }) {
-  const [product] = useGetItem(id);
-  return (
-    <Container>
-      <ScrollToTop />
-      <CardProduct product={product} />
-    </Container>
-  );
-}
-
-export default DeatilsMod;
