+++ pages/products/[id].jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import useRouterQuery from "../../hooks/useRouterQuery";
-import DeatilsMod from "../../modules/products/details";
+import DeatilsMod from "../../modules/products";
 
 const ProductDetails = () => {
   const id = useRouterQuery();
